feat: makes album-library optional
