+++ src/app/(app)/layout.tsx
@@ -16,17 +16,15 @@ const AppLayout = async ({ children, albums, signin }: Props) => {
 
   return (
     <Providers>
-      <div className='mobile:flex-col-reverse flex h-full flex-row justify-center overflow-hidden *:w-full '>
-        <div className='mobile:gap-4 big:max-w-[90ch] flex flex-col justify-between gap-8 *:px-[var(--window-padding-x)] first:*:pt-[var(--window-padding-y)] last:*:pb-[var(--window-padding-y)]'>
+      <div className='flex h-full flex-row justify-center overflow-hidden mobile:flex-col-reverse '>
+        <div className='flex w-full flex-col justify-between gap-8 *:px-[var(--window-padding-x)] first:*:pt-[var(--window-padding-y)] last:*:pb-[var(--window-padding-y)] mobile:gap-4 big:max-w-[90ch]'>
           <main className='flex h-full flex-col justify-between gap-4'>
             {children}
           </main>
           <Footer />
         </div>
-        <aside className='h-full overflow-scroll px-[var(--window-padding-x)] py-[var(--window-padding-y)]'>
-          {session && albums}
-          {!session && signin}
-        </aside>
+        {session && albums}
+        {!session && signin}
       </div>
     </Providers>
   );
